fix(vim): use `normal!` command to jump to last open position instead of simulating keystroke
When the file is modified outside of neovim, a 'BufRead' event will be
triggered if the user enters the buffer containing that file. Normally
this is fine, but when the buffer is opened in the preview window of
Snacks picker, the simulated keystrokes 'g`"' mess up with neovim's mode
transition. So this patch uses `normal!` to replace the original hacky
method.