fix(shrun): fix a bug that causes separator being displayed even if there is only one task
To reproduce the bug, first make sure there are two tasks in sidebar,
delete the top task, and then reopen sidebar. Now there is a separator
below the only task. The root cause is: after 038af64 disables output
tail, when there is only one task, the buffer line count of sidebar is
1 instead of 2, thus when calling `display_panel`, `render_sidebar_from_scratch`
will be called, and since task id may be greater than 1, the separator
is drawed. To fix this issue, we should use the actual count of tasks to
decide if separator is needed.

Fixes: d2ddf89a "refactor(shrun): use `desc_sorted_pairs` to iterate table by key's order"